Using Passport.js to Add Cookies and Sessions
Now the website is able to remember when a user is already logged in so that once they log-in and they want to be able to access the parts of the website that require authentication they don't have to keep logging back in. This is because it's all saved back in our cookie for that particular session and the content of the cookie has a meaning to our server because it can check against it and know that the current user is already signed in so we don't have to force them to sign in again when they try to access a part of the website that requires authentication.